pyproject.toml
[project]
name = "recipe-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = [
    "django>=5.1.5",
    "djangorestframework-simplejwt>=5.4.0",
    "djangorestframework>=3.15.2",
    "ruff>=0.9.2",
    "celery>=5.4.0",
    "django-debug-toolbar>=5.0.1",
    "django-filter>=24.3",
    "django-redis>=5.4.0",
    "django-silk>=5.3.2",
    "djoser>=2.3.1",
    "drf-nested-routers>=0.94.1",
    "pillow>=11.1.0",
    "psycopg[binary,pool]",
    "redis>=5.2.1",
    "pytest>=8.3.4",
    "argon2-cffi>=23.1.0",
    "drf-spectacular>=0.28.0",
    "flake8>=7.1.1",
    "gunicorn>=23.0.0",
    "django-cors-headers>=4.6.0"
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
    "pytest-django>=4.9.0",
    "pytest-watch>=4.2.0",
]
